move shared code
