+++ package.json
@@ -10,7 +10,7 @@
     "build": "tsc",
     "clean": "rm -rf dist cdk.out",
     "watch": "tsc -w",
-    "test": "jest --detectOpenHandles --forceExit --testPathIgnorePatterns=integ/ dist/",
+    "test": "jest --detectOpenHandles --forceExit --testPathIgnorePatterns=integ/ dist/ cdk.out/",
     "integ-test": "jest --detectOpenHandles --runInBand --testPathPattern=integ/",
     "cdk": "cdk",
     "fix": "run-s fix:*",
